Refactor Hello/Greeting into hello.rs
