+++ epp-client/src/tests/se.rs
@@ -27,6 +27,7 @@ mod request {
     use crate::domain::update::DomainAddRemove;
     use crate::domain::update::DomainChangeInfo;
     use crate::domain::update::EppDomainUpdate;
+    use crate::hello::EppHello;
     use crate::host::check::EppHostCheck;
     use crate::host::create::EppHostCreate;
     use crate::host::delete::EppHostDelete;
@@ -36,7 +37,7 @@ mod request {
     use crate::host::update::HostChangeInfo;
     use crate::message::ack::EppMessageAck;
     use crate::message::poll::EppMessagePoll;
-    use crate::request::{EppHello, EppLogin, EppLogout};
+    use crate::request::{EppLogin, EppLogout};
     use crate::xml::EppXml;
     use chrono::{DateTime, NaiveDate};
     use std::str::FromStr;
